Update to support python3 and drop python2
This is in advance of the Django 2.0 upgrade, at which
point python2 will no longer be officially supported. The
guidance for packaged app developers is that they should
mirror the Django releases - so this branch represents
that switch. It does not support 2+3 together - once we
switch to this branch we will be Django 2.0 / Python 3.6
only.